requirements.txt
pymongo
python-dotenv
langchain_mongodb